Basic testNatsToSparkConnector JUnit Test
